requirements.txt
# ── AI / LLM ────────────────────────────────────────────────────────────────────
anthropic>=0.40.0

# ── Notion ──────────────────────────────────────────────────────────────────────
notion-client>=2.2.1

# ── HTTP ────────────────────────────────────────────────────────────────────────
httpx>=0.27.0

# ── Data validation ─────────────────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.3.0

# ── Environment ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.1

# ── Google Drive (migration only) ───────────────────────────────────────────────
google-api-python-client>=2.130.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# ── Async ───────────────────────────────────────────────────────────────────────
anyio>=4.4.0

# ── Logging ─────────────────────────────────────────────────────────────────────
structlog>=24.1.0
rich>=13.7.0

# ── Testing ─────────────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0
respx>=0.21.0

# ── Dev tools ───────────────────────────────────────────────────────────────────
black>=24.4.0
isort>=5.13.0
mypy>=1.10.0
